Cist: Ant. de B.M.V. corrected
The [Ant 1] in C11 and C12 should be shorter in Cist. Rite, the same goes for Suffrage of Our Lady between Trinity and Advent in Vespers.

This fixes the situation for Latin and Czech, I haven't seen the (rubrica cisterciense) tags in other languages.

BTW, I have noticed one more thing: in the Suffrage of Our Lady between Trinity and Advent in Vespers, the Collect is ended with the short conclusion instead of `$Per eundem`. And if I replace the `$Per eumdem Christum` with `$Per eumdem`, it shows double Amen, and maybe breaks something else (post partum?), so I left it be for now...